move skip plot wrapper to samples_plotter.py
