+++ autofit/non_linear/nest/dynesty/plotter.py
@@ -1,51 +1,7 @@
 from dynesty import plotting as dyplot
-from functools import wraps
-import os
 
-from autoconf import conf
 from autofit.plot import SamplesPlotter
-
-
-
-def skip_plot_in_test_mode(func):
-    """
-    Skips visualization plots of non-linear searches if test mode is on.
-
-    Parameters
-    ----------
-    func
-        A function which plots a result of a non-linear search.
-
-    Returns
-    -------
-        A function that plots a visual, or None if test mode is on.
-    """
-
-    @wraps(func)
-    def wrapper(
-        obj: object,
-        *args,
-        **kwargs
-    ):
-        """
-        Skips visualization plots of non-linear searches if test mode is on.
-
-        Parameters
-        ----------
-        obj
-            An plotter object which performs visualization of a non-linear search.
-
-        Returns
-        -------
-            A function that plots a visual, or None if test mode is on.
-        """
-
-        if "PYAUTOFIT_TEST_MODE" in os.environ:
-            return
-
-        return func(obj, *args, **kwargs)
-
-    return wrapper
+from autofit.plot.samples_plotters import skip_plot_in_test_mode
 
 
 class DynestyPlotter(SamplesPlotter):
